Predictive neural guider, PHD2-style guide graph, dithering fix
Neural guide model:
- Outcome-based training: hindsight-optimal targets from observed next-frame
  error replace P-controller mimicry, enabling the model to learn PE patterns
- Multi-scale features (22 inputs): 4-frame history + short/medium/long means
  (10f/60f/300f) + accumulated gear error (traces PE curve) + mount state
- Blend ramp-in: 0% → 50% over 480 frames (~16 min), like PHD2's learning phase
- Experience buffer 512 → 2048 entries (~68 min at 2s cadence, multiple PE cycles)

Guide graph improvements:
- Correction pulse bars (log-scaled, blue RA / orange Dec) extending up/down
  from zero line — shows what the controller is doing, not just the error
- Dither markers (yellow dashed vertical lines) at dither events
- Settling shading (dim yellow tint) during post-dither settle periods
- Y scale from visible window samples (ignores settling spikes)
- Wider sample spacing (6px vs 3px) for readability
- Guider state label: settling distance, correction arrows, paused during slew

Dithering fix:
- DitherAsync now offsets the guide star lock position (was no-op before)
- Guide stats polled during settling (was gated on isGuiding only)
- Calibration overlay hides once guiding starts

Other:
- IMountDriver.GetAxisPositionAsync returns long? (was int?)
- GuideStats exposes LastRaPulseMs/LastDecPulseMs for correction display
- CLAUDE.md: fix stale OpenGL refs, add AppData paths, update query keys
- TODO.md: add flaky CI tests, CLI offline training command

+++ src/TianWen.Lib/Devices/Fake/FakeMountDriver.cs
@@ -325,14 +325,14 @@ internal sealed class FakeMountDriver(FakeDevice fakeDevice, IExternal external)
     // Simulates encoder ticks: 360° = EncoderTicksPerRevolution ticks
     private const int EncoderTicksPerRevolution = 11_520_000; // typical high-res encoder
 
-    public async ValueTask<int?> GetAxisPositionAsync(TelescopeAxis axis, CancellationToken cancellationToken)
+    public async ValueTask<long?> GetAxisPositionAsync(TelescopeAxis axis, CancellationToken cancellationToken)
     {
         using var @lock = await _sem.AcquireLockAsync(cancellationToken);
         UpdateTrackingState();
         return axis switch
         {
-            TelescopeAxis.Primary => (int)(ConditionRA(_ra + _accumulatedRaHours) / 24.0 * EncoderTicksPerRevolution),
-            TelescopeAxis.Seconary => (int)((_dec + _accumulatedDecDegrees + 90.0) / 360.0 * EncoderTicksPerRevolution),
+            TelescopeAxis.Primary => (long)(ConditionRA(_ra + _accumulatedRaHours) / 24.0 * EncoderTicksPerRevolution),
+            TelescopeAxis.Seconary => (long)((_dec + _accumulatedDecDegrees + 90.0) / 360.0 * EncoderTicksPerRevolution),
             _ => null
         };
     }
